Upgrade to PySide6

+++ src/itaxotools/common/utility.py
@@ -17,35 +17,98 @@
 #-----------------------------------------------------------------------------
 
 
-"""Utility functions for PyQt5"""
+"""Utility functions for PySide6"""
 
-import PyQt5.QtCore as QtCore
-import PyQt5.QtWidgets as QtWidgets
-import PyQt5.QtGui as QtGui
+from PySide6 import QtCore
+from PySide6 import QtWidgets
+from PySide6 import QtStateMachine
+from PySide6 import QtGui
 
-import logging
+from contextlib import contextmanager
+
 import sys, os, io
 import multiprocessing
 
+
 ##############################################################################
-### Logging
+### IO redirection
 
-class StdioLogger():
-    """Redirect system stdout and stderr to logger"""
-    def __init__(self, logger=None):
-        if logger is None:
-            logger = logging.getLogger()
-        self.logger = logger
+@contextmanager
+def _redirect(module=sys, stream='stdout', dest=None):
+    """Redirect module stream to file stream"""
+    original = getattr(module, stream)
+    original.flush()
+    setattr(module, stream, dest)
+    try:
+        yield dest
+    finally:
+        dest.flush()
+        setattr(module, stream, original)
 
-    def __enter__(self):
-        self.err = sys.stderr.write
-        self.out = sys.stdout.write
-        sys.stderr.write = self.logger.error
-        sys.stdout.write = self.logger.info
+@contextmanager
+def redirect(module=sys, stream='stdout', dest=None, mode='w'):
+    """
+    Redirect module's stream according to `dest`:
+    - If None: Do nothing
+    - If String: Open file and redirect
+    - Else: Assume IOWrapper, redirect
+    """
+    if dest is None:
+        yield getattr(module, stream)
+    elif isinstance(dest, str):
+        with open(dest, mode) as file, _redirect(module, stream, file) as f:
+            yield f
+    else:
+        with _redirect(module, stream, dest) as f:
+            yield f
 
-    def __exit__(self, et, ev, tr):
-        sys.stderr.write = self.err
-        sys.stdout.write = self.out
+
+class TextEditLogger(QtWidgets.QPlainTextEdit):
+    """Thread-safe log display in a QPlainTextEdit"""
+    appendRecord = QtCore.Signal(object)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.setReadOnly(True)
+        self.appendRecord.connect(self.appendTextInline)
+
+    @QtCore.Slot(object)
+    def appendTextInline(self, text):
+        self.moveCursor(QtGui.QTextCursor.End);
+        self.insertPlainText(text);
+        self.moveCursor(QtGui.QTextCursor.End);
+
+    def append(self, text):
+        self.appendRecord.emit(str(text))
+
+class TextEditIO(io.IOBase):
+    """File-like object that writes to TextEditLogger"""
+
+    def __init__(self, widget):
+        super().__init__()
+        self.widget = widget
+
+    def close(self):
+        pass
+
+    def flush(self):
+        pass
+
+    def readable(self):
+        return False
+
+    def writeable(self):
+        return True
+
+    def write(self, text):
+        self.widget.append(text)
+
+    def writeline(self, line):
+        self.write(line+'\n')
+
+    def writelines(self, lines):
+        for line in lines:
+            self.writeline(line)
 
 class PipeIO(io.IOBase):
     """File-like object that writes to a pipe connection"""
@@ -129,32 +192,13 @@ class PipeIO(io.IOBase):
             self.connection.send(self.buffer)
         self.buffer = ''
 
-class TextEditLogger(QtWidgets.QPlainTextEdit):
-    """Thread-safe log display in a QPlainTextEdit"""
-    appendRecord = QtCore.pyqtSignal(object)
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.setReadOnly(True)
-        self.handler = logging.Handler()
-        self.handler.emit = self.emit
-        self.appendRecord.connect(self.appendTextInline)
-
-    def appendTextInline(self, text):
-        self.moveCursor(QtGui.QTextCursor.End);
-        self.insertPlainText(text);
-        self.moveCursor(QtGui.QTextCursor.End);
-
-    def emit(self, record):
-        text = self.handler.format(record)
-        self.appendRecord.emit(text)
 
 ##############################################################################
 ### Multiprocessing
 
 class UThread(QtCore.QThread):
     """
-    Multithreaded function execution using PyQt5.
+    Multithreaded function execution using PySide6.
     Initialise with the function and parameters for execution.
     Signals are used to communicate with parent thread.
     Use self.start() to fork/spawn.
@@ -171,8 +215,8 @@ class UThread(QtCore.QThread):
     fail(exception):
         Emitted if an exception occured. Passes the exception.
     """
-    done = QtCore.pyqtSignal(object)
-    fail = QtCore.pyqtSignal(object)
+    done = QtCore.Signal(object)
+    fail = QtCore.Signal(object)
 
     def __init__(self, function, *args, **kwargs):
         """
@@ -204,91 +248,23 @@ class UThread(QtCore.QThread):
             self.done.emit(result)
 
 
-class UProcess(QtCore.QThread):
-    """
-    Multiprocess function execution using PyQt5.
-    Unlike QtCore.QProcess, this launches python functions, not programs.
-    Use self.start() to fork/spawn.
-
-    Example
-    ----------
-
-    def work(number):
-        print('This runs on the child process.')
-        print('Trying to use PyQt in here won\'t work.')
-        return number * 2
-
-    def success(result):
-        print('This runs on the parent process.')
-        print('You can spawn dialog messages here.')
-        QMessageBox.information(None, 'Success',
-            'Result = '+str(result), QMessageBox.Ok)
-        print('This prints 42: ', result)
-
-    self.process = UProcess(work, 21)
-    self.process.done.connect(success)
-    self.process.start()
-    return
-
-    """
-    done = QtCore.pyqtSignal(object)
-    fail = QtCore.pyqtSignal(object)
+class UWorker():
 
     def __getstate__(self):
         """Required for process spawning."""
         state = self.__dict__.copy()
-        state['logger'] = None
-        state['handleOut'] = self._loggerNone
-        state['handleErr'] = self._loggerNone
         return state
 
     def __setstate__(self, state):
         """Required for process spawning."""
-        # super(UProcess, self).__init__(None)
         self.__dict__ = state
-        return
 
-    def __init__(self, function, *args, **kwargs):
-        """
-        Parameters
-        ----------
-        function : function
-            The function to run on the new process.
-        *args : positional argument, optional
-            If given, it is passed on to the function.
-        **kwargs : keyword arguments, optional
-            If given, it is passed on to the function.
-        """
-        super().__init__()
-        self.logger = None
-        self.pipeControl = multiprocessing.Pipe(duplex=True)
-        self.pipeData = multiprocessing.Pipe(duplex=True)
-        self.pipeOut = multiprocessing.Pipe(duplex=False)
-        self.pipeErr = multiprocessing.Pipe(duplex=False)
-        self.pipeIn = multiprocessing.Pipe(duplex=False)
-        self.process = multiprocessing.Process(
-            target=self.target, daemon=True,
-            args=(function,)+args, kwargs=kwargs)
-
-    def setLogger(self, logger):
-        """Send process output to given logger"""
-        self.logger = logger
-        # return
-        if logger is not None:
-            self.handleOut = self._loggerOut
-            self.handleErr = self._loggerErr
-        else:
-            self.handleOut = self._loggerNone
-            self.handleErr = self._loggerNone
-
-    def _loggerNone(self, data):
-        pass
-
-    def _loggerOut(self, data):
-        self.logger.info(data)
-
-    def _loggerErr(self, data):
-        self.logger.error(data)
+    def __init__(self, dict):
+        self.pipeControl = dict['pipeControl']
+        self.pipeData = dict['pipeData']
+        self.pipeOut = dict['pipeOut']
+        self.pipeErr = dict['pipeErr']
+        self.pipeIn = dict['pipeIn']
 
     def target(self, function, *args, **kwargs):
         """
@@ -320,6 +296,77 @@ class UProcess(QtCore.QThread):
         except Exception as exception:
             self.pipeControl.send('EXCEPTION')
             self.pipeData.send(exception)
+
+class UProcess(QtCore.QThread):
+    """
+    Multiprocess function execution using PySide6.
+    Unlike QtCore.QProcess, this launches python functions, not programs.
+    Use self.start() to fork/spawn.
+
+    Example
+    ----------
+
+    def work(number):
+        print('This runs on the child process.')
+        print('Trying to use PyQt in here won\'t work.')
+        return number * 2
+
+    def success(result):
+        print('This runs on the parent process.')
+        print('You can spawn dialog messages here.')
+        QMessageBox.information(None, 'Success',
+            'Result = '+str(result), QMessageBox.Ok)
+        print('This prints 42: ', result)
+
+    self.process = UProcess(work, 21)
+    self.process.done.connect(success)
+    self.process.start()
+    return
+
+    """
+    done = QtCore.Signal(object)
+    fail = QtCore.Signal(object)
+
+    def __init__(self, function, *args, **kwargs):
+        """
+        Parameters
+        ----------
+        function : function
+            The function to run on the new process.
+        *args : positional argument, optional
+            If given, it is passed on to the function.
+        **kwargs : keyword arguments, optional
+            If given, it is passed on to the function.
+        """
+        super().__init__()
+        self.stream = None
+        self.pipeControl = multiprocessing.Pipe(duplex=True)
+        self.pipeData = multiprocessing.Pipe(duplex=True)
+        self.pipeOut = multiprocessing.Pipe(duplex=False)
+        self.pipeErr = multiprocessing.Pipe(duplex=False)
+        self.pipeIn = multiprocessing.Pipe(duplex=False)
+        self.worker = UWorker(self.__dict__)
+        self.process = multiprocessing.Process(
+            target=self.worker.target, daemon=True,
+            args=(function,)+args, kwargs=kwargs)
+
+    def setStream(self, stream):
+        """Send process output to given file-like stream"""
+        self.stream = stream
+        # return
+        if stream is not None:
+            self.handleOut = self._streamOut
+            self.handleErr = self._streamOut
+        else:
+            self.handleOut = self._loggerNone
+            self.handleErr = self._loggerNone
+
+    def _loggerNone(self, data):
+        pass
+
+    def _streamOut(self, data):
+        self.stream.write(data)
+
 
     def run(self):
         """
@@ -374,7 +421,8 @@ class UProcess(QtCore.QThread):
 
     def quit(self):
         """Clean exit"""
-        self.process.terminate()
+        if self.process is not None and self.process.is_alive():
+            self.process.terminate()
         super().quit()
 
 ##############################################################################
@@ -382,15 +430,18 @@ class UProcess(QtCore.QThread):
 
 class NamedEvent(QtCore.QEvent):
     """Custom event for use in state machines"""
-    userEvent = QtCore.QEvent.User + 1042
+    userEvent = QtCore.QEvent.registerEventType()
+    events = set()
     def __init__(self, name, *args, **kwargs):
         """Pass name and args"""
-        super().__init__(self.userEvent)
+        super().__init__(QtCore.QEvent.Type(self.userEvent))
         self.name = name
         self.args = args
         self.kwargs = kwargs
+        # Avoid garbage-collection
+        NamedEvent.events.add(self)
 
-class NamedTransition(QtCore.QAbstractTransition):
+class NamedTransition(QtStateMachine.QAbstractTransition):
     """Custom transition for use in state machines"""
     def __init__(self, name):
         """Only catch events with given name"""
@@ -403,4 +454,5 @@ class NamedTransition(QtCore.QAbstractTransition):
         return False
     def onTransition(self, event):
         """Override virtual function"""
-        pass
+        # Allow event to be garbage-collected
+        # NamedEvent.events.remove(event)
